+++ android/app/src/main/java/com/teameetmeet/meetmeet/data/network/api/AuthApi.kt
@@ -1,5 +1,6 @@
 package com.teameetmeet.meetmeet.data.network.api
 
+import com.teameetmeet.meetmeet.data.network.entity.AccessTokenRequest
 import com.teameetmeet.meetmeet.data.network.entity.LoginResponse
 import com.teameetmeet.meetmeet.data.network.entity.RefreshAccessTokenRequest
 import retrofit2.http.Body
@@ -9,4 +10,7 @@ interface AuthApi {
 
     @POST("auth/refresh")
     suspend fun refreshAccessToken(@Body refreshAccessTokenRequest: RefreshAccessTokenRequest): LoginResponse
+
+    @POST("auth/check/token")
+    suspend fun checkValidAccessToken(@Body accessTokenRequest: AccessTokenRequest)
 }
