+++ android/app/src/main/java/com/teameetmeet/meetmeet/data/repository/TokenRepository.kt
@@ -1,11 +1,18 @@
 package com.teameetmeet.meetmeet.data.repository
 
+import android.util.Log
 import com.teameetmeet.meetmeet.data.ExpiredRefreshTokenException
 import com.teameetmeet.meetmeet.data.NoDataException
 import com.teameetmeet.meetmeet.data.local.datastore.DataStoreHelper
 import com.teameetmeet.meetmeet.data.network.api.AuthApi
+import com.teameetmeet.meetmeet.data.network.entity.AccessTokenRequest
 import com.teameetmeet.meetmeet.data.network.entity.RefreshAccessTokenRequest
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.flow.map
+import retrofit2.HttpException
 import javax.inject.Inject
 
 class TokenRepository @Inject constructor(
@@ -19,9 +26,27 @@ class TokenRepository @Inject constructor(
             dataStoreHelper.storeAppToken(response.accessToken, response.refreshToken)
             return response.accessToken
         } catch (e: Exception) {
-            println(e)
             throw ExpiredRefreshTokenException()
         }
+    }
+
+    fun autoLoginApp(token: String): Flow<Unit> {
+        return flowOf(true)
+            .map {
+                authApi.checkValidAccessToken(AccessTokenRequest(token))
+            }.catch {
+                when(it) {
+                    is HttpException -> {
+                        if(it.code() == 418) {
+                            val token = refreshAccessToken()
+                            autoLoginApp(token)
+                        } else {
+                            throw it
+                        }
+                    }
+                    else -> throw it
+                }
+            }
     }
 
     suspend fun getAccessToken(): String {
